# PERF: __getitem__