This is in proof-of-concept phase at the moment.

Goals:
1) improve performance of `Series.__getitem__`, `DataFrame.iat`, etc.
2) Make `_get_dict_of_block_index` clearer, use some of the terminology/patterns used in pandas' BlockManager for the analogous lookups

Perf comparisons `iat` vs a hypothetical implementation currently parked at `iat2`, and `Series.__getitem__`
```
import numpy as np
import modin.pandas as pd

arr = np.random.randn(20_000, 50)
df = pd.DataFrame(arr)

x = 19000
y = 40

%timeit df.iat[x, y]
4.04 ms ± 526 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1.55 ms ± 9.43 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- iat2
11.6 µs ± 417 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- pandas

ser = df[10]
_ = ser._query_compiler.to_pandas()  # not sure why this is needed; discussed below

%timeit ser[50]
3.69 ms ± 241 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
1.31 ms ± 23 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
1.91 µs ± 43.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- pandas
```

Without the `ser._query_compiler.to_pandas()` call, `ser[50]` incorrectly returns `arr[50, 0]` instead of `arr[50, 10]`.  I have  no idea what's going on there.