@pyrito 2022-08-05
@jbrockmendel I think the `to_pandas()` call flushes out the queue so any lazy operations will get evaluated.