@jbrockmendel 2022-08-09
>  I think the to_pandas() call flushes out the queue so any lazy operations will get evaluated.

is there a maximally-cheap way to do this?