@mvashishtha 2022-08-09
> is there a maximally-cheap way to do this?

@jbrockmendel what do you mean by "maximally cheap"? 

We do have #2814 tracking improvements to `to_pandas`.