@jbrockmendel 2022-08-09
> what do you mean by "maximally cheap"?

is there a way to flush the queue that is cheaper than to_pandas?