@mvashishtha 2022-08-09
> is there a way to flush the queue that is cheaper than to_pandas?

We usually use `BenchmarkMode.put(True)` for this purpose. That should drain call queues and make everything that's usually async finish immediately: 

https://github.com/modin-project/modin/blob/ef654c414364ca319812114324ad9d85e83ee68e/modin/core/dataframe/pandas/partitioning/partition_manager.py#L64-L75

But on the other hand, because we lose the ability to be async, it might not reflect practical performance accurately.