@jbrockmendel 2022-08-09
I meant something to do inside Series._getitem to get the accurate results