@mvashishtha 2022-08-09
> I meant something to do inside Series._getitem to get the accurate results

If `df` is made of `M by N` partitions, `df[10]` creates `M` new partitions object that point to the same physical data as the first column of `df`'s partitions. These `M` partitions each have something like `lambda x: x.iloc[:, 10]` on their call queue. My guess is that the remote code you've added is skipping the call queue, so you select from column 0 instead of column 10 of the original frame. `to_pandas` will flush the call queue so that `_data` of the partitions of `df[10]` contains only column 10. If you need to get a remote function result from a partition, you have to go through something like `apply`, which adds the function to the existing call queue.

ref. my comment from our earlier discussion https://github.com/modin-project/modin/issues/4729#issuecomment-1199858851