@jbrockmendel 2022-08-10
So... what would be an edit to add to this PR in `Series._getitem` to get correct results?